+++ module/apps/activity-schedule.js
@@ -384,8 +384,8 @@ export class ActivitySchedule extends HandlebarsApplicationMixin(ApplicationV2) 
     return data;
   }
 
-  _onRender(context, options) {
-    super._onRender(context, options);
+  async _onRender(context, options) {
+    await super._onRender(context, options);
     const html = this.element;
     html.querySelector(".change-year").addEventListener("change", this._setYear.bind(this));