+++ module/apps/aura-config.js
@@ -66,8 +66,8 @@ export class AuraConfig extends HandlebarsApplicationMixin(ApplicationV2) {
     return context;
   }
 
-  _onRender(context, options) {
-    super._onRender(context, options);
+  async _onRender(context, options) {
+    await super._onRender(context, options);
     const html = this.element;
     html.querySelectorAll("input").forEach((input) => {
       input.addEventListener("focusin", (ev) => ev.target.select());